requirements.txt
flask
sentence-transformers
chromadb
langchain
transformers
huggingface_hubx`
torch
# langchain-community
